feat: renamed package to make more clear its purpose
