+++ github/go.mod
@@ -1,22 +1,23 @@
-module github.com/GoogleCloudPlatform/deploystack/dsgithub
+module github.com/GoogleCloudPlatform/deploystack/github
 
 go 1.20
 
 require (
-	github.com/GoogleCloudPlatform/deploystack v0.0.0-20230211225259-7dd3e51a73f5
-	github.com/GoogleCloudPlatform/deploystack/gcloudtf v0.0.0-20230211225259-7dd3e51a73f5
+	github.com/GoogleCloudPlatform/deploystack v0.0.0-20230219100415-70850fce76e7
+	github.com/GoogleCloudPlatform/deploystack/terraform v0.0.0-20230219100415-70850fce76e7
 	gopkg.in/src-d/go-git.v4 v4.13.1
 )
 
 require (
-	cloud.google.com/go v0.109.0 // indirect
+	cloud.google.com/go v0.110.0 // indirect
 	cloud.google.com/go/compute v1.18.0 // indirect
 	cloud.google.com/go/compute/metadata v0.2.3 // indirect
 	cloud.google.com/go/domains v0.8.0 // indirect
-	cloud.google.com/go/iam v0.10.0 // indirect
-	cloud.google.com/go/longrunning v0.4.0 // indirect
+	cloud.google.com/go/iam v0.12.0 // indirect
+	cloud.google.com/go/longrunning v0.4.1 // indirect
 	cloud.google.com/go/scheduler v1.8.0 // indirect
 	cloud.google.com/go/storage v1.29.0 // indirect
+	github.com/GoogleCloudPlatform/deploystack/gcloud v0.0.0-20230216051911-d4fae17aa4a8 // indirect
 	github.com/Microsoft/go-winio v0.6.0 // indirect
 	github.com/agext/levenshtein v1.2.3 // indirect
 	github.com/apparentlymart/go-textseg/v13 v13.0.0 // indirect
@@ -25,32 +26,31 @@ require (
 	github.com/golang/protobuf v1.5.2 // indirect
 	github.com/google/go-cmp v0.5.9 // indirect
 	github.com/google/uuid v1.3.0 // indirect
-	github.com/googleapis/enterprise-certificate-proxy v0.2.2 // indirect
+	github.com/googleapis/enterprise-certificate-proxy v0.2.3 // indirect
 	github.com/googleapis/gax-go/v2 v2.7.0 // indirect
 	github.com/hashicorp/hcl v1.0.0 // indirect
-	github.com/hashicorp/hcl/v2 v2.14.1 // indirect
-	github.com/hashicorp/terraform-config-inspect v0.0.0-20221020162138-81db043ad408 // indirect
+	github.com/hashicorp/hcl/v2 v2.16.1 // indirect
+	github.com/hashicorp/terraform-config-inspect v0.0.0-20230201191712-8cad743c8c26 // indirect
 	github.com/jbenet/go-context v0.0.0-20150711004518-d14ea06fba99 // indirect
 	github.com/kevinburke/ssh_config v1.2.0 // indirect
 	github.com/mitchellh/go-homedir v1.1.0 // indirect
 	github.com/mitchellh/go-wordwrap v1.0.1 // indirect
-	github.com/nyaruka/phonenumbers v1.1.5 // indirect
 	github.com/sergi/go-diff v1.3.1 // indirect
 	github.com/src-d/gcfg v1.4.0 // indirect
 	github.com/xanzy/ssh-agent v0.3.3 // indirect
-	github.com/zclconf/go-cty v1.12.0 // indirect
+	github.com/zclconf/go-cty v1.12.1 // indirect
 	go.opencensus.io v0.24.0 // indirect
 	golang.org/x/crypto v0.6.0 // indirect
 	golang.org/x/mod v0.8.0 // indirect
-	golang.org/x/net v0.6.0 // indirect
+	golang.org/x/net v0.7.0 // indirect
 	golang.org/x/oauth2 v0.5.0 // indirect
 	golang.org/x/sys v0.5.0 // indirect
 	golang.org/x/text v0.7.0 // indirect
 	golang.org/x/tools v0.6.0 // indirect
 	golang.org/x/xerrors v0.0.0-20220907171357-04be3eba64a2 // indirect
-	google.golang.org/api v0.109.0 // indirect
+	google.golang.org/api v0.110.0 // indirect
 	google.golang.org/appengine v1.6.7 // indirect
-	google.golang.org/genproto v0.0.0-20230209215440-0dfe4f8abfcc // indirect
+	google.golang.org/genproto v0.0.0-20230216225411-c8e22ba71e44 // indirect
 	google.golang.org/grpc v1.53.0 // indirect
 	google.golang.org/protobuf v1.28.1 // indirect
 	gopkg.in/src-d/go-billy.v4 v4.3.2 // indirect
