+++ go.work
@@ -2,7 +2,7 @@ go 1.19
 
 use (
 	.
-	./dsgithub
+	./github
 	./dstester
 	./gcloud
 	./terraform
